chore(mempool_node): move component config
commit-id:03ddb09f
