+++ crates/mempool_node/src/config/mod.rs
@@ -1,6 +1,7 @@
 #[cfg(test)]
 mod config_test;
 
+pub mod component_config;
 pub mod component_execution_config;
 
 use std::collections::BTreeMap;
@@ -9,6 +10,7 @@ use std::path::Path;
 use std::sync::LazyLock;
 
 use clap::Command;
+use component_config::*;
 use component_execution_config::*;
 use papyrus_config::dumping::{append_sub_config_name, ser_pointer_target_param, SerializeConfig};
 use papyrus_config::loading::load_and_process_config;
@@ -21,7 +23,7 @@ use starknet_consensus_manager::config::ConsensusManagerConfig;
 use starknet_gateway::config::{GatewayConfig, RpcStateReaderConfig};
 use starknet_http_server::config::HttpServerConfig;
 use starknet_sierra_compile::config::SierraToCasmCompilationConfig;
-use validator::{Validate, ValidationError};
+use validator::Validate;
 
 use crate::version::VERSION_FULL;
 
@@ -40,67 +42,6 @@ pub static CONFIG_POINTERS: LazyLock<ConfigPointers> = LazyLock::new(|| {
         ],
     )]
 });
-
-/// The components configuration.
-#[derive(Clone, Debug, Serialize, Deserialize, Validate, PartialEq)]
-#[validate(schema(function = "validate_components_config"))]
-pub struct ComponentConfig {
-    #[validate]
-    pub batcher: ComponentExecutionConfig,
-    #[validate]
-    pub consensus_manager: ComponentExecutionConfig,
-    #[validate]
-    pub gateway: ComponentExecutionConfig,
-    #[validate]
-    pub http_server: ComponentExecutionConfig,
-    #[validate]
-    pub mempool: ComponentExecutionConfig,
-}
-
-impl Default for ComponentConfig {
-    fn default() -> Self {
-        Self {
-            batcher: ComponentExecutionConfig::batcher_default_config(),
-            consensus_manager: ComponentExecutionConfig::consensus_manager_default_config(),
-            gateway: ComponentExecutionConfig::gateway_default_config(),
-            http_server: ComponentExecutionConfig::http_server_default_config(),
-            mempool: ComponentExecutionConfig::mempool_default_config(),
-        }
-    }
-}
-
-impl SerializeConfig for ComponentConfig {
-    fn dump(&self) -> BTreeMap<ParamPath, SerializedParam> {
-        #[allow(unused_mut)]
-        let mut sub_configs = vec![
-            append_sub_config_name(self.batcher.dump(), "batcher"),
-            append_sub_config_name(self.consensus_manager.dump(), "consensus_manager"),
-            append_sub_config_name(self.gateway.dump(), "gateway"),
-            append_sub_config_name(self.http_server.dump(), "http_server"),
-            append_sub_config_name(self.mempool.dump(), "mempool"),
-        ];
-
-        sub_configs.into_iter().flatten().collect()
-    }
-}
-
-pub fn validate_components_config(components: &ComponentConfig) -> Result<(), ValidationError> {
-    // TODO(Tsabary/Lev): We need to come up with a better mechanism for this validation, simply
-    // listing all components and expecting one to remember adding a new component to this list does
-    // not suffice.
-    if components.gateway.execute
-        || components.mempool.execute
-        || components.batcher.execute
-        || components.http_server.execute
-        || components.consensus_manager.execute
-    {
-        return Ok(());
-    }
-
-    let mut error = ValidationError::new("Invalid components configuration.");
-    error.message = Some("At least one component should be allowed to execute.".into());
-    Err(error)
-}
 
 /// The configurations of the various components of the node.
 #[derive(Debug, Deserialize, Serialize, Clone, PartialEq, Validate)]
